git-svn-id: http://posttool.com/svn/posttool/PSWebV1.0@3903 241bf0be-db3f-0410-a781-e906b5fc445c

+++ src/com/pagesociety/web/module/resource/ResourceModule.java
@@ -578,7 +578,8 @@ public class ResourceModule extends WebStoreModule
 			{
 				/* if we cant create a resource we need to delete the upload */
 				try{
-					path_provider.delete(path_token);
+					if(!update)
+						path_provider.delete(path_token);
 				}catch(Exception e){e.printStackTrace();}//swallow this one
 				
 				ERROR(pe);
